backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=1.4.0
pydantic==2.5.2
pydantic-settings>=2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic>=1.7.0
pytest==7.4.3
httpx>=0.23.0
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow>=2.8.0
torch>=1.9.0
transformers>=4.11.0
aiohttp==3.9.1
websockets==12.0
bcrypt==4.0.1
aiofiles==23.2.1
tenacity==8.2.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
joblib==1.3.2
requests==2.31.0
